bump up es client to v8 & golang to 1.24
