+++ Makefile.Common
@@ -15,7 +15,12 @@ misspell:
 lint: misspell
 	$(LINT) run --allow-parallel-runners
 
+.PHONY: toolchain
+toolchain:
+	$(GOCMD) get toolchain@none
+
 .PHONY: tidy
 tidy:
 	rm -fr go.sum
-	$(GOCMD) mod tidy -compat=1.22.0
+	$(GOCMD) mod tidy -compat=1.24.0
+	make toolchain
